server/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==5.0.3
django-cors-headers==4.4.0
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-nested-routers==0.94.1
exceptiongroup==1.2.2
executing==2.1.0
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.27.0
jedi==0.19.1
kombu==5.3.6
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.3.3
pytest-django==4.9.0
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
watchdog==5.0.2
wcwidth==0.2.13
